Solution to the given assignment
Since I was not allowed to use any other module, I had to use pygame to get input from the webcam, but this feature was experimental and as a result, did not run on Windows. Hence, I installed a virtual machine and worked on it. But I was getting some errors in getting access to the webcam from the ubuntu within the virtual machine, due to time constraints I was not able to resolve the issue. In order to make up for it, my code takes a bunch of photographs and processes it, like it would have done with the incoming frames from the webcam. It is done such that only a few changes need to be done in order to take the input from the webcam.